requirements.txt
# Core dependencies
playwright==1.40.0
python-dotenv==1.0.0

# Data handling
pandas==2.1.4
openpyxl==3.1.2

# Utilities
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
PyYAML==6.0.1

# Logging and monitoring
loguru==0.7.2

# AI/ML
openai==1.54.0

# PDF Processing
pypdf==4.0.1

# Scheduling
APScheduler==3.10.4

# Security
cryptography==42.0.0
